+++ Sources/Diagnostics/DiagnosticsTracker.swift
@@ -43,7 +43,8 @@ protocol DiagnosticsTrackerType: Sendable {
                                    responseCode: Int,
                                    backendErrorCode: Int?,
                                    resultOrigin: HTTPResponseOrigin?,
-                                   verificationResult: VerificationResult,
+                                   verificationResult: SignatureVerificationResult,
+                                   responseRequestDate: Date?,
                                    isRetry: Bool,
                                    connectionErrorReason: ConnectionErrorReason?)
 
@@ -245,22 +246,29 @@ final class DiagnosticsTracker: DiagnosticsTrackerType, Sendable {
                                    responseCode: Int,
                                    backendErrorCode: Int?,
                                    resultOrigin: HTTPResponseOrigin?,
-                                   verificationResult: VerificationResult,
+                                   verificationResult: SignatureVerificationResult,
+                                   responseRequestDate: Date?,
                                    isRetry: Bool,
                                    connectionErrorReason: ConnectionErrorReason?) {
-        self.trackEvent(name: .httpRequestPerformed,
-                        properties: DiagnosticsEvent.Properties(
-                            verificationResult: verificationResult.name,
-                            endpointName: endpointName,
-                            host: host,
-                            responseTime: responseTime,
-                            successful: wasSuccessful,
-                            responseCode: responseCode,
-                            backendErrorCode: backendErrorCode,
-                            etagHit: resultOrigin == .cache,
-                            isRetry: isRetry,
-                            connectionErrorReason: connectionErrorReason
-                        ))
+        self.trackEvent(
+            name: .httpRequestPerformed,
+            properties: DiagnosticsEvent.Properties(
+                verificationResult: verificationResult.result.name,
+                verificationFailureReason: verificationResult.failureReason?.rawValue,
+                verificationDeviceClockOffsetMinutes: responseRequestDate.map {
+                    Int(self.dateProvider.now().timeIntervalSince($0) / 60)
+                },
+                endpointName: endpointName,
+                host: host,
+                responseTime: responseTime,
+                successful: wasSuccessful,
+                responseCode: responseCode,
+                backendErrorCode: backendErrorCode,
+                etagHit: resultOrigin == .cache,
+                isRetry: isRetry,
+                connectionErrorReason: connectionErrorReason
+            )
+        )
     }
 
     func trackPurchaseAttempt(wasSuccessful: Bool,